+++ src/sql/workbench/contrib/resourceViewer/browser/resourceViewer.contribution.ts
@@ -8,6 +8,20 @@ import { Registry } from 'vs/platform/registry/common/platform';
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors';
 import { ResourceViewerEditor } from 'sql/workbench/contrib/resourceViewer/browser/resourceViewerEditor';
 import { ResourceViewerInput } from 'sql/workbench/browser/editor/resourceViewer/resourceViewerInput';
+import { CommandsRegistry } from 'vs/platform/commands/common/commands';
+import { ServicesAccessor, IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { IEditorService, ACTIVE_GROUP } from 'vs/workbench/services/editor/common/editorService';
+
+CommandsRegistry.registerCommand({
+	id: 'resourceViewer.newResourceViewer',
+	handler: (accessor: ServicesAccessor, ...args: any[]): void => {
+		const instantiationService: IInstantiationService = accessor.get(IInstantiationService);
+		const editorService: IEditorService = accessor.get(IEditorService);
+
+		const resourceViewerInput = instantiationService.createInstance(ResourceViewerInput);
+		editorService.openEditor(resourceViewerInput, { pinned: true }, ACTIVE_GROUP);
+	}
+});
 
 const resourceViewerDescriptor = EditorDescriptor.create(
 	ResourceViewerEditor,
